CrossPlatformLightweightServer/requirements.txt
google_cloud_speech
google_cloud_storage
pyaudio
pyinstaller
